fix(pipeline): isolate LangGraph config adapter from graph wiring
- Move RunnableConfig boundary to langgraph_node_adapter.py so graph.py
  satisfies tests/types RunnableConfig import guard.
- Use Optional[RunnableConfig] from langchain_core.runnables.base so LangGraph
  injects runtime config (Union X | None is not on its whitelist).
